Added Gravity Type Option
Choose between:
-realistic (takes dist into account)
-pretty (doesn't take dist into account, but makes neat orbits)
Default color range is now prettier

+++ FunWithParticles/SliderBar.cpp
@@ -9,7 +9,7 @@ SliderBar::SliderBar(){
 	isHeld = false;
 }
 
-SliderBar::SliderBar(int x, int y, int width, ALLEGRO_COLOR color, int min, int max, int startPos){
+SliderBar::SliderBar(int x, int y, int width, ALLEGRO_COLOR color, int min, int max, float startPos){
 	this->min = min;
 	this->max = max;
 	sliderX = startPos;
